requirements.txt
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-pagedown==1.0.6
gunicorn==19.9.0
markdown2==2.3.7
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.1.2
